Add check and create directory for currency rates file

+++ main.py
@@ -11,6 +11,10 @@ console = Console()
 
 def prepare_currency():
     output_path = "./data/currency_rates.csv"
+    if os.path.exists(output_path):
+        return output_path
+    root_path = os.path.dirname(output_path)
+    os.makedirs(root_path, exist_ok=True)
     country_currency = CurrencyRate(path="./configs/countries_currency.csv")
     country_currency = country_currency.get_country_currency()
     country_currency.to_csv(output_path, index=False)
